Use AbstractMetricMessageHandler as MetricMessageHandler
