+++ server/sink/src/main/java/org/bithon/server/sink/event/MetricOverEventHandler.java
@@ -18,7 +18,7 @@ package org.bithon.server.sink.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.bithon.server.sink.metrics.AbstractMetricMessageHandler;
+import org.bithon.server.sink.metrics.MetricMessageHandler;
 import org.bithon.server.sink.metrics.transformer.TopoTransformers;
 import org.bithon.server.storage.datasource.DataSourceSchemaManager;
 import org.bithon.server.storage.datasource.input.IInputRow;
@@ -34,7 +34,7 @@ import java.io.IOException;
  * @date 2022/8/2 22:30
  */
 @Slf4j
-public class MetricOverEventHandler extends AbstractMetricMessageHandler implements EventMessageHandler<IInputRow> {
+public class MetricOverEventHandler extends MetricMessageHandler implements EventMessageHandler<IInputRow> {
 
     private final String eventType;
     private final ObjectMapper objectMapper;
@@ -50,7 +50,8 @@ public class MetricOverEventHandler extends AbstractMetricMessageHandler impleme
               topoTransformers,
               metaStorage,
               metricStorage,
-              dataSourceSchemaManager);
+              dataSourceSchemaManager,
+              null);
         this.eventType = eventType;
         this.objectMapper = objectMapper;
     }
